refactor: :truck: naming improvements - rename filenames and components to align with the name in the UI more
